+++ git_well/main.py
@@ -15,10 +15,12 @@ class GitWellModalCLI(scfg.ModalCLI):
     from git_well.git_rebase_add_continue import __cli__ as rebase_add_continue
     from git_well.git_remote_protocol import __cli__ as remote_protocol
     from git_well.git_discover_remote import __cli__ as discover_remote
+    from git_well.git_autoconf_gpgsign import __cli__ as autoconf_gpgsign
 
 
 def main():
-    modal = GitWellModalCLI()
+    from git_well import __version__
+    modal = GitWellModalCLI(version=__version__)
     modal.main()